% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_to_directus.R
\name{auth_to_directus}
\alias{auth_to_directus}
\title{Authentification into the API}
\usage{
auth_to_directus(base.url, login, password, directus_version = 9)
}
\arguments{
\item{base.url}{A string. A project name}

\item{login}{A string.}

\item{password}{A string.}

\item{directus_version}{A number.}
}
\value{
Message about success auth
}
\description{
Authentification into the API
}
