% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directusInstance.R
\name{directusInstance}
\alias{directusInstance}
\title{R6 Class REST API of Directus}
\description{
Initialization and working with Directus Headless CMS
}
\details{
A person can also greet you.
}
\examples{

## ------------------------------------------------
## Method `directusInstance$auth`
## ------------------------------------------------

db<-initDirectus(db='palebot')
db$auth(login="YOUR LOGIN", password="YOUR PASSWORD")
#' # expected:
# you are authentificated to..

## ------------------------------------------------
## Method `directusInstance$get_items`
## ------------------------------------------------

db<-initDirectus(db='palebot')
db$auth(login="YOUR LOGIN", password="YOUR PASSWORD")
items<-db$get_items('tablename', params="?limit=10")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{db}}{The name of \href{https://docs.directus.io/getting-started/glossary.html#projects}{Directus project}}

\item{\code{base.url}}{Base url}

\item{\code{db.info}}{db.info}

\item{\code{current.table}}{current.table}

\item{\code{auth.status}}{is authentificated}

\item{\code{directus.version}}{version of directus}

\item{\code{table.names}}{list of tables}

\item{\code{durls}}{urls}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-directusInstance-new}{\code{directusInstance$new()}}
\item \href{#method-directusInstance-print}{\code{directusInstance$print()}}
\item \href{#method-directusInstance-auth}{\code{directusInstance$auth()}}
\item \href{#method-directusInstance-get_items}{\code{directusInstance$get_items()}}
\item \href{#method-directusInstance-clone}{\code{directusInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-directusInstance-new"></a>}}
\if{latex}{\out{\hypertarget{method-directusInstance-new}{}}}
\subsection{Method \code{new()}}{
Create a new directus API instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{directusInstance$new(db, directus.version = 9, db.info = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{a string. The name of \href{https://docs.directus.io/getting-started/glossary.html#projects}{Directus project}}

\item{\code{directus.version}}{Version of Directus.}

\item{\code{db.info}}{H list with information about path to Directus and available tables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{directusInstance} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-directusInstance-print"></a>}}
\if{latex}{\out{\hypertarget{method-directusInstance-print}{}}}
\subsection{Method \code{print()}}{
print some information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{directusInstance$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-directusInstance-auth"></a>}}
\if{latex}{\out{\hypertarget{method-directusInstance-auth}{}}}
\subsection{Method \code{auth()}}{
Auth
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{directusInstance$auth(login, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{login}}{a string.}

\item{\code{password}}{a string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
success message.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{db<-initDirectus(db='palebot')
db$auth(login="YOUR LOGIN", password="YOUR PASSWORD")
#' # expected:
# you are authentificated to..
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-directusInstance-get_items"></a>}}
\if{latex}{\out{\hypertarget{method-directusInstance-get_items}{}}}
\subsection{Method \code{get_items()}}{
Get Items
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{directusInstance$get_items(tablename, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tablename}}{a string.}

\item{\code{params}}{a string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dataframe with items.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{db<-initDirectus(db='palebot')
db$auth(login="YOUR LOGIN", password="YOUR PASSWORD")
items<-db$get_items('tablename', params="?limit=10")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-directusInstance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-directusInstance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{directusInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
