% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initDirectus.R
\name{initDirectus}
\alias{initDirectus}
\title{Initialization and working with \href{https://docs.directus.io/reference/introduction.html}{Directus REST API}}
\usage{
initDirectus(db, directus.version = 9, db.info = NA)
}
\arguments{
\item{db}{A string. A project name}

\item{directus.version}{A number.}

\item{db.info}{A list with information about path to Directus and available tables.}
}
\value{
An instance of directusInstance class
}
\description{
Initialization and working with \href{https://docs.directus.io/reference/introduction.html}{Directus REST API}
}
\details{
\subsection{Methods}{
\subsection{Authentification}{

For authenticating into the API run \code{db$auth()}with parameters.

\if{html}{\out{<div class="sourceCode">}}\preformatted{db$auth(login="YOUR LOGIN", password="YOUR PASSWORD")
# expected:
# you are authentificated to..
}\if{html}{\out{</div>}}
}

\subsection{Get Items}{

Run \code{db$get_items()} with parameters.
}

\subsection{Code example:}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{items<-db$get_items('tablename', params="?limit=10")
# expected:
# dataframe with items
}\if{html}{\out{</div>}}
}

}
}
\examples{
library(directusR)
db1<-initDirectus(db='paleobot')
db1
# or
db2<-initDirectus(db='paleosib', directus.version=8)
db2
db.info <- list(
  arth=list(
   base.url='http://arth.mironcat.tk/',
   tbl.names=c('specimens','collections','localities','taxa','reflinks')
  ),
  myproject = list(
   base.url='http://url.to.myproject/',
   tbl.names=c('table1','table2')
  )
 )
db3<-initDirectus(db='myproject', db.info=db.info)
db3
}
\seealso{
\code{\link[=directusInstance]{directusInstance()}} - class documentation
}
